+++ processor/src/test/resources/expected/com/example/arez/ComponentWithEventHandler_.java
@@ -108,7 +108,8 @@ class ComponentWithEventHandler_ extends ComponentWithEventHandler {
 
     void componentWillUnmount();
 
-    boolean shouldComponentUpdate(@Nonnull BaseProps arg0, @Nonnull BaseState arg1, @Nonnull BaseContext arg2);
+    boolean shouldComponentUpdate(@Nonnull BaseProps arg0, @Nonnull BaseState arg1,
+        @Nonnull BaseContext arg2);
   }
 
   static final class NativeReactComponent extends NativeAdapterComponent<BaseProps, BaseState, BaseContext, ComponentWithEventHandler> implements Lifecycle {
@@ -137,7 +138,8 @@ class ComponentWithEventHandler_ extends ComponentWithEventHandler {
     }
 
     @Override
-    public boolean shouldComponentUpdate(@Nonnull final BaseProps arg0, @Nonnull final BaseState arg1, @Nonnull final BaseContext arg2) {
+    public boolean shouldComponentUpdate(@Nonnull final BaseProps arg0,
+        @Nonnull final BaseState arg1, @Nonnull final BaseContext arg2) {
       return performShouldComponentUpdate(arg0,arg1,arg2);
     }
   }