+++ processor/src/test/resources/expected/com/example/arez/ComponentWithNoAutoActionEventHandler_.java
@@ -76,7 +76,8 @@ class ComponentWithNoAutoActionEventHandler_ extends ComponentWithNoAutoActionEv
 
     void componentWillUnmount();
 
-    boolean shouldComponentUpdate(@Nonnull BaseProps arg0, @Nonnull BaseState arg1, @Nonnull BaseContext arg2);
+    boolean shouldComponentUpdate(@Nonnull BaseProps arg0, @Nonnull BaseState arg1,
+        @Nonnull BaseContext arg2);
   }
 
   static final class NativeReactComponent extends NativeAdapterComponent<BaseProps, BaseState, BaseContext, ComponentWithNoAutoActionEventHandler> implements Lifecycle {
@@ -105,7 +106,8 @@ class ComponentWithNoAutoActionEventHandler_ extends ComponentWithNoAutoActionEv
     }
 
     @Override
-    public boolean shouldComponentUpdate(@Nonnull final BaseProps arg0, @Nonnull final BaseState arg1, @Nonnull final BaseContext arg2) {
+    public boolean shouldComponentUpdate(@Nonnull final BaseProps arg0,
+        @Nonnull final BaseState arg1, @Nonnull final BaseContext arg2) {
       return performShouldComponentUpdate(arg0,arg1,arg2);
     }
   }